+++ packages/vue/src/index.ts
@@ -2,4 +2,8 @@ export { useScribeSite } from "./useScribeSite.js";
 export type { UseScribeSiteResult } from "./useScribeSite.js";
 export { useScribeArticle } from "./useScribeArticle.js";
 export type { UseScribeArticleResult } from "./useScribeArticle.js";
+export { useScribePublicationUri } from "./useScribePublicationUri.js";
+export type { UseScribePublicationUriResult } from "./useScribePublicationUri.js";
+export { useScribeDocumentUri } from "./useScribeDocumentUri.js";
+export type { UseScribeDocumentUriResult } from "./useScribeDocumentUri.js";
 export type { Site, Article, ArticleRef, SiteGroup } from "@scribe-atp/core";